pyproject.toml
[tool.poetry]
name = "slh-sh"
version = "0.1.0"
description = ""
authors = ["Nova Sky <2989791+0xnovasky@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.scripts]
slh-sh = "slh_sh.main:app"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
typer = { extras = ["all"], version = "^0.9.0" }
pyyaml = "^6.0.1"
pandas = "^2.1.1"
requests = "^2.31.0"
bs4 = "^0.0.1"
pdfminer-six = "^20221105"
gspread = "^5.11.2"
oauth2client = "^4.1.3"
pymupdf = "^1.23.3"
scholarly = "^1.7.11"
sqlalchemy = "^2.0.21"
pydantic = "^2.4.0"
alembic = "^1.12.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
